refactor(dashboard): consolidate ISO week utils into date-utils.ts
Closes #165

Both getCurrentIsoWeek() and parseIsoWeekBounds() were duplicated inline
in PatternsPage.tsx and WeekSelector.tsx respectively, each with a "keep
in sync" comment referencing the server implementation. Extract both into
dashboard/src/lib/date-utils.ts with a single top-of-file sync comment.

No logic changes — pure consolidation.
