+++ dashboard/src/components/patterns/WeekSelector.tsx
@@ -1,34 +1,13 @@
 import { useEffect } from 'react';
 import { ChevronLeft, ChevronRight } from 'lucide-react';
 import { Button } from '@/components/ui/button';
+import { parseIsoWeekBounds } from '@/lib/date-utils';
 import type { WeekInfo } from '@/lib/api';
 
 interface WeekSelectorProps {
   currentWeek: string;      // e.g., "2026-W10"
   weeks: WeekInfo[];        // from GET /api/reflect/weeks, most recent first
   onWeekChange: (week: string) => void;
-}
-
-// Parse an ISO week string into UTC Monday/Sunday boundaries.
-// Adapted from parseIsoWeek in server/src/routes/shared-aggregation.ts --
-// uses inclusive end (Sunday) for display instead of exclusive end (next Monday) for SQL queries.
-// Kept here to avoid a server-side import in the dashboard bundle.
-function parseIsoWeekBounds(weekStr: string): { start: Date; end: Date } | null {
-  const match = /^(\d{4})-W(\d{2})$/.exec(weekStr);
-  if (!match) return null;
-
-  const year = parseInt(match[1], 10);
-  const week = parseInt(match[2], 10);
-
-  const jan4 = new Date(Date.UTC(year, 0, 4));
-  const jan4Day = jan4.getUTCDay();
-  const daysToMonday = jan4Day === 0 ? 6 : jan4Day - 1;
-  const week1Monday = new Date(jan4.getTime() - daysToMonday * 86400000);
-
-  const start = new Date(week1Monday.getTime() + (week - 1) * 7 * 86400000);
-  const end = new Date(start.getTime() + 6 * 86400000); // Sunday (inclusive for display)
-
-  return { start, end };
 }
 
 // Format a UTC date as "Mar 4" using UTC to avoid local timezone day shift.
