+++ dashboard/src/pages/PatternsPage.tsx
@@ -7,6 +7,7 @@ import { WeekSelector } from '@/components/patterns/WeekSelector';
 import { WeekAtAGlanceStrip } from '@/components/patterns/WeekAtAGlanceStrip';
 import { CollapsibleCategoryList } from '@/components/patterns/CollapsibleCategoryList';
 import { WorkingStyleHighlights } from '@/components/patterns/WorkingStyleHighlights';
+import { getCurrentIsoWeek } from '@/lib/date-utils';
 import { parseSSEStream } from '@/lib/sse';
 import { Card, CardContent, CardHeader, CardTitle, CardDescription } from '@/components/ui/card';
 import { Button } from '@/components/ui/button';
@@ -29,30 +30,6 @@ function formatRelativeDate(iso: string): string {
   if (hours < 24) return `${hours}h ago`;
   const days = Math.floor(hours / 24);
   return `${days}d ago`;
-}
-
-// Compute the current ISO week identifier (YYYY-WNN) in UTC.
-// Mirrors formatIsoWeek/parseIsoWeek in server/src/routes/shared-aggregation.ts
-// -- kept here to avoid a server-side import in the dashboard bundle.
-// IMPORTANT: keep in sync with the canonical server implementation.
-function getCurrentIsoWeek(): string {
-  const now = new Date();
-  const nowDay = now.getUTCDay();
-  const daysToMonday = nowDay === 0 ? 6 : nowDay - 1;
-  const monday = new Date(now.getTime() - daysToMonday * 86400000);
-
-  // Thursday of this week determines the ISO year
-  const thursday = new Date(monday.getTime() + 3 * 86400000);
-  const year = thursday.getUTCFullYear();
-
-  // Find Monday of week 1 for this ISO year
-  const jan4 = new Date(Date.UTC(year, 0, 4));
-  const jan4Day = jan4.getUTCDay();
-  const daysToW1Monday = jan4Day === 0 ? 6 : jan4Day - 1;
-  const week1Monday = new Date(jan4.getTime() - daysToW1Monday * 86400000);
-
-  const weekNum = Math.round((monday.getTime() - week1Monday.getTime()) / (7 * 86400000)) + 1;
-  return `${year}-W${String(weekNum).padStart(2, '0')}`;
 }
 
 export default function PatternsPage() {
